pyproject.toml
[project]
name = "my-media-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
]
